sign up and login route created
Landing page, Signup and Login Route are Created through Next.js and data is saved in database ( PostgreSQL ). When sign up your user email will be saved in database and next time you can login with your email that is saved in our database. Now i am creating a next page in which we will be directed when we log in